Move the existing middleware into a Laravel namespace
